added workflow basics, deployed onto render, add repo environment varible to send render hook

+++ server/src/seeds/cleanDb.ts
@@ -1,11 +1,14 @@
 import models from '../models/index.js';
 import db from '../config/connection.js';
 
-export default async (modelName: "Question", collectionName: string) => {
+export default async (modelName: "Question" , collectionName: string) => {
   try {
+    if (!models[modelName].db.db) {
+      throw new Error(`Model or database for ${modelName} does not exist`);
+    }
     let modelExists = await models[modelName].db.db.listCollections({
       name: collectionName
-    }).toArray()
+    }).toArray();
 
     if (modelExists.length) {
       await db.dropCollection(collectionName);
